iwlwifi: remove lq_sta->last_rate_n_flags
This variable is supposed to represent the last/current transmitted rate, and
is used in rs_get_rate() as if it corresponds to the rate for
lq_sta->last_txrate_idx. However, these two variables are not set synchronously
and this seems to be the cause of some "Heisenbugs". Thus, remove this variable
and instead use the properties of the active rate scaling table to set the
variables that last_rate_n_flags was used for.

Signed-off-by: Daniel C Halperin <daniel.c.halperin@intel.com>
Signed-off-by: Reinette Chatre <reinette.chatre@intel.com>