iwlwifi: put card to sleep when unassociated/idle
When mac80211 signals that the card is idle, we can instruct
the microcode to sleep. mac80211 will wake it up for the
association period, and after associating will instruct us
to go to PS mode (which we do again, but mac80211 doesn't
know that we use PS mode for idle).

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Signed-off-by: Reinette Chatre <reinette.chatre@intel.com>