iwlwifi: validate the signature for EEPROM and OTP
Both 1000 & 6000 series NIC contains on-chip OTP memory that
replaces the off-chip EEPROM memory. The nature of OTP means
there is a limited number of times a particular board can go through the
factory flow and be (re)calibrated. There will be some number of boards
that contain EEPROM memory because OTP blocks are full.

That means driver need to support both configuration 1) OTP only 2) OTP
and EEPROM.

To support these configuration, driver needs to check both signatures.
If the OTP signature is not valid and the valid EEPROM signature is found,
then EEPROM memory should be used.

Signed-off-by: Wey-Yi Guy <wey-yi.w.guy@intel.com>
Signed-off-by: Reinette Chatre <reinette.chatre@intel.com>