iwlwifi: clean up last_phy_res
The last_phy_res[100] variable is used in an odd
way. The first byte of it is used as a flag, which
however is missing being set to 0 again. Also, the
array need only be 61 bytes, but really is just a
flag and a struct iwl_rx_phy_res.

Clean this up by splitting the variable into two:
last_phy_res and last_phy_res_valid, using correct
types for both (struct and bool).

Additionally, the valid variable needs to be reset
to false when the data is being used. However, it
can happen that two PHY data replies are received
after another, at least in the case that the frame
overflowed the available FIFO space.

While doing all this also move the variable to the
_agn part of the hw-specific union.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Reinette Chatre <reinette.chatre@intel.com>