iwlwifi: use newer firmware format for everything *but* 4965
patch 'iwlwifi: modify hardware check' introduced an
error in determining which firmware format to used for 4965. the
comment indicates that for 4965, it should fall through the case but
the comparison only selects the 4965 to break out.  this makes all
newer (legacy not the new new tlv formatted) ucode files fail to load
since with the error size too small.

change is to exit when not 4965 and allow the 4965 to fall through as
the comment indicates.

Signed-off-by: Jay Sternberg <jay.e.sternberg@intel.com>
Signed-off-by: Reinette Chatre <reinette.chatre@intel.com>